# a7_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'EatZy Booking Web API'
  description: >-
    Web Resources Specification (A7) for EatZy Booking.
    Includes Web UI Routes (HTTP 302 Redirects) and REST API Routes (JSON).
...
paths:
...
  /login:
    get:
      operationId: R101
      summary: 'R101: Login Form'
      description: 'Provide login form. Access: VIS'
      tags:
        - 'M01: Authentication'
      responses:
        '200':
          description: 'Ok. Show login UI'
    post:
      operationId: R102
      summary: 'R102: Login Action'
      description: 'Processes the login form submission. Access: VIS'
      tags:
        - 'M01: Authentication'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required:
                - email
                - password
      responses:
        '302':
          description: 'Redirect after processing login.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  value: '/restaurants'
                302Error:
                  value: '/login'
...
  /restaurants:
    get:
      operationId: R201
      summary: 'R201: Browse Restaurants'
      description: 'List restaurants with search filters. Access: VIS'
      tags:
        - 'M02: Restaurants'
      parameters:
        - in: query
          name: search
          description: Full text search
          schema:
            type: string
        - in: query
          name: capacity
          schema:
            type: integer
      responses:
        '200':
          description: 'Ok. Show list UI.'
...
  /api/restaurants:
    get:
      operationId: API_R211
      summary: 'Search Restaurants API'
      description: 'Returns JSON array of restaurants. Access: PUB'
      tags:
        - 'M02: Restaurants'
      parameters:
        - in: query
          name: query
          description: Full text search string
          schema:
            type: string
        - in: query
          name: capacity
          description: Filter by minimum capacity
          schema:
            type: integer
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    rating:
                      type: number
...