# app/Http/Controllers/TwoFactorController.php
<?php

namespace App\Http\Controllers;

use Illuminate\Http\Request;
use Illuminate\Support\Facades\Auth;
use App\Models\User;

class TwoFactorController extends Controller
{
    public function showSetup()
    {
        $user = Auth::user();
        $google2fa = new \PragmaRX\Google2FA\Google2FA();

        if (!$user->two_factor_secret) {
            $secret = $google2fa->generateSecretKey();
            $user->two_factor_secret = $secret;
            $user->save();
        } else {
            $secret = $user->two_factor_secret;
        }

        $qrCodeUrl = $google2fa->getQRCodeUrl(
            config('app.name', 'EatZ'),
            $user->email,
            $secret
        );

        $renderer = new \BaconQrCode\Renderer\ImageRenderer(
            new \BaconQrCode\Renderer\RendererStyle\RendererStyle(200),
            new \BaconQrCode\Renderer\Image\SvgImageBackEnd()
        );
        $writer = new \BaconQrCode\Writer($renderer);
        $qrCodeSvg = $writer->writeString($qrCodeUrl);

        return view('auth.2fa.setup', compact('secret', 'qrCodeSvg', 'user'));
    }

    public function enable(Request $request)
    {
        $request->validate([
            'code' => 'required|string|size:6',
        ]);

        $user = Auth::user();

        if (!$user->verify2FACode($request->code)) {
            return back()->withErrors(['code' => 'Invalid verification code.']);
        }

        $user->update(['two_factor_enabled' => \DB::raw('true')]);

        return redirect()->route('account.me')
            ->with('success', 'Two-factor authentication enabled successfully.');
    }

    public function showDisable()
    {
        return view('auth.2fa.disable');
    }

    public function disable(Request $request)
    {
        $request->validate([
            'password' => 'required',
        ]);

        $user = Auth::user();

        if (!\Hash::check($request->password, $user->password)) {
            return back()->withErrors(['password' => 'Incorrect password.']);
        }

        $user->two_factor_secret = null;
        $user->update(['two_factor_enabled' => \DB::raw('false')]);

        return redirect()->route('account.me')
            ->with('success', 'Two-factor authentication disabled.');
    }

    public function showVerify()
    {
        if (!session()->has('2fa:user_id')) {
            return redirect()->route('login');
        }

        return view('auth.2fa.verify');
    }

    public function verify(Request $request)
    {
        $request->validate([
            'code' => 'required|string|size:6',
        ]);

        $userId = session('2fa:user_id');
        if (!$userId) {
            return redirect()->route('login');
        }

        $user = User::find($userId);
        if (!$user) {
            return redirect()->route('login');
        }

        if (!$user->verify2FACode($request->code)) {
            return back()->withErrors(['code' => 'Invalid verification code.']);
        }

        session()->forget('2fa:user_id');
        Auth::login($user);
        $request->session()->regenerate();

        return redirect()->intended('/restaurants');
    }
}